Move util::vec::byte_swap to canvas_traits
